export types outside the Karpenter addon
